+++ template/wlcontroller/c/eigenc.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -23,7 +24,7 @@ extern "C" {
 // index for col-major access
 #define Cind(n, i, j) ((i) + (j)*n)
 
-void matMult(float *C, const float *A, const float *B, size_t m, size_t n, size_t k, float alpha, bool AT, bool BT);
+void matMult(float *C, const float *A, const float *B, int64_t m, int64_t n, int64_t k, float alpha, bool AT, bool BT);
 
 #ifdef __cplusplus
 }
